move associated files helper to utils
